+++ src/config/mod.rs
@@ -1,7 +1,7 @@
 use crate::cli::Cli;
 use miette::Diagnostic;
 use owo_colors::OwoColorize;
-use std::fs::{DirEntry, File};
+use std::fs::File;
 use std::io::{self, stdin, IsTerminal, Read};
 use std::path::{Path, PathBuf};
 use std::{env, fs};
@@ -19,14 +19,8 @@ pub struct PathAndLineCount {
     pub line_count: usize,
 }
 
-pub struct FolderInfo {
-    pub folder_name: PathBuf,
-    pub file_paths: Vec<PathBuf>,
-}
-
 pub enum Input {
     File(PathAndLineCount),
-    Folder(FolderInfo),
     Command(String),
     Stdin,
 }
@@ -41,18 +35,19 @@ pub enum Output {
 pub enum ConfigError {
     #[error("Missing filename ({0} for help)")]
     MissingFilename(String),
+
     #[error("Cannot read from both file and {0}")]
     CannotReadBothFileAndListenCommand(String),
+
     #[error("Could not determine input type")]
     CouldNotDetermineInputType,
+
     #[error("{0}: No such file or directory")]
     NoSuchFileOrDirectory(String),
-    #[error("Path is neither a file nor a directory")]
-    PathNotFileNorDirectory,
-    #[error("Path is not a directory")]
-    PathNotDirectory,
-    #[error("Unable to read directory")]
-    UnableToReadDirectory,
+
+    #[error("Path is not a file")]
+    PathNotFile,
+
     #[error("I/O Error: {0}")]
     Io(#[from] io::Error),
 }
@@ -62,13 +57,13 @@ pub fn create_config(args: &Cli) -> Result<Config, ConfigError> {
 
     validate_input(
         has_data_from_stdin,
-        args.file_or_folder_path.is_some(),
+        args.file_path.is_some(),
         args.listen_command.is_some(),
     )?;
 
     let input = get_input(args, has_data_from_stdin)?;
     let output = get_output(has_data_from_stdin, args.to_stdout);
-    let follow = should_follow(args.follow, args.listen_command.is_some(), &input);
+    let follow = should_follow(args.follow, args.listen_command.is_some());
 
     Ok(Config {
         input,
@@ -80,14 +75,14 @@ pub fn create_config(args: &Cli) -> Result<Config, ConfigError> {
 
 fn validate_input(
     has_data_from_stdin: bool,
-    has_file_or_folder_input: bool,
+    has_file_input: bool,
     has_follow_command_input: bool,
 ) -> Result<(), ConfigError> {
-    if !has_data_from_stdin && !has_file_or_folder_input && !has_follow_command_input {
+    if !has_data_from_stdin && !has_file_input && !has_follow_command_input {
         return Err(ConfigError::MissingFilename("tspin --help".magenta().to_string()));
     }
 
-    if has_file_or_folder_input && has_follow_command_input {
+    if has_file_input && has_follow_command_input {
         return Err(ConfigError::CannotReadBothFileAndListenCommand(
             "--listen-command".magenta().to_string(),
         ));
@@ -101,8 +96,8 @@ fn get_input(args: &Cli, has_data_from_stdin: bool) -> Result<Input, ConfigError
         Ok(Input::Stdin)
     } else if let Some(command) = &args.listen_command {
         Ok(Input::Command(command.clone()))
-    } else if let Some(path) = &args.file_or_folder_path {
-        let path = PathBuf::from(path);
+    } else if let Some(path_str) = &args.file_path {
+        let path = PathBuf::from(path_str);
         process_path_input(path)
     } else {
         Err(ConfigError::CouldNotDetermineInputType)
@@ -120,67 +115,22 @@ fn get_output(has_data_from_stdin: bool, to_stdout: bool) -> Output {
 }
 
 fn process_path_input(path: PathBuf) -> Result<Input, ConfigError> {
-    match get_path_type(&path)? {
-        PathType::File => {
-            let line_count = count_lines(&path)?;
-            Ok(Input::File(PathAndLineCount { path, line_count }))
-        }
-        PathType::Folder => {
-            let mut file_paths = list_files_in_directory(&path)?;
-            file_paths.sort();
-            Ok(Input::Folder(FolderInfo {
-                folder_name: path,
-                file_paths,
-            }))
-        }
-    }
-}
-
-enum PathType {
-    File,
-    Folder,
-}
-
-fn get_path_type(path: &Path) -> Result<PathType, ConfigError> {
-    let metadata = fs::metadata(path).map_err(|_| ConfigError::NoSuchFileOrDirectory(path.display().to_string()))?;
-    if metadata.is_file() {
-        Ok(PathType::File)
-    } else if metadata.is_dir() {
-        Ok(PathType::Folder)
-    } else {
-        Err(ConfigError::PathNotFileNorDirectory)
-    }
-}
-
-const fn should_follow(follow_flag: bool, has_command: bool, input: &Input) -> bool {
-    if has_command || matches!(input, Input::Folder(_)) {
-        true
-    } else {
-        follow_flag
-    }
-}
-
-fn list_files_in_directory(path: &Path) -> Result<Vec<PathBuf>, ConfigError> {
-    if !path.is_dir() {
-        return Err(ConfigError::PathNotDirectory);
+    if !path.exists() {
+        return Err(ConfigError::NoSuchFileOrDirectory(path.display().to_string()));
     }
 
-    fs::read_dir(path)
-        .map_err(|_| ConfigError::UnableToReadDirectory)?
-        .filter_map(Result::ok)
-        .filter(is_normal_file)
-        .map(|entry| Ok(entry.path()))
-        .collect()
+    let metadata = fs::metadata(&path)?;
+    if !metadata.is_file() {
+        return Err(ConfigError::PathNotFile);
+    }
+
+    let line_count = count_lines(&path)?;
+
+    Ok(Input::File(PathAndLineCount { path, line_count }))
 }
 
-fn is_normal_file(entry: &DirEntry) -> bool {
-    entry.path().is_file()
-        && entry
-            .path()
-            .file_name()
-            .and_then(|name| name.to_str())
-            .map(|name| !name.starts_with('.'))
-            .unwrap_or(false)
+const fn should_follow(follow_flag: bool, has_command: bool) -> bool {
+    if has_command { true } else { follow_flag }
 }
 
 fn count_lines<P: AsRef<Path>>(file_path: P) -> Result<usize, ConfigError> {